Tap into the Observable
The HeroService methods will tap into the flow of observable values and send a message, using the log() method we defined

Note: The RxJS tap() operator is a utility operator used to intercept each emission on the source observable and runs a function but returns an output that is identical to the source observable as long as it doesn't find any error.
The operator is generally used for debugging observables for correct values or performing other side effects.

Syntax: tap(observer, error, complete): Observable

Parameter Explanation
observer or nextOrObserver: This specifies a normal observer object or a callback for next. This is the same as the source observable. It is an optional argument. Its default value is undefined.
error: It is used to specify the error method if any error occurs in the source. It is also an optional argument. Its default value is undefined.
complete: This is called when the task is complete. It is also an optional argument. Its default value is undefined.