Add HeroService.updateHero() to HeroService class
The overall structure of updateHero() is similar to getHeroes(), but it uses http.put() to persist the changed hero on the server

Note: The HttpClient.put() method takes 3 parameters:
1. The URL
2. The data to update (modified hero)
3. Options

The URL is unchanged. The heroes web API knows which hero to update by looking at the hero's id

The heroes web API expects a special header in the HTTP save requests. That header is in the httpOptions constant which we will add to HeroService.